Minor suggested refactoring & expand int type constraints
